requirements.txt
Django==2.0.7
django-filter==2.0.0
djangorestframework==3.9.0
Markdown==3.0.1
Pillow==5.2.0
pytz==2018.5
